fix project not building on system with non-default libcxx
I dunno what about systems where everything (or at least clang) is
built against libcxx, but when it just installed, for clang to use its
includes, flag -stdlib=libc++ must be provided.

If we do this, we can make use of things like std::variant, which are
yet to appear in standard standard libraries.